+++ tests/disjoint_set_vec.rs
@@ -1,6 +1,6 @@
-use disjoint::DisjointSet;
+use disjoint::DisjointSetVec;
 
-fn verify_subsets(disjoint_set: &DisjointSet, expected_subsets: &[Vec<usize>]) {
+fn verify_subsets<T>(disjoint_set: &DisjointSetVec<T>, expected_subsets: &[Vec<usize>]) {
     assert_eq!(
         disjoint_set.len(),
         expected_subsets.iter().map(|subset| subset.len()).sum()
@@ -22,7 +22,7 @@ fn verify_subsets(disjoint_set: &DisjointSet, expected_subsets: &[Vec<usize>]) {
 
 #[test]
 fn constructor_prodcues_singletons() {
-    let disjoint_set = DisjointSet::new(5);
+    let disjoint_set = DisjointSetVec::from(vec!["a", "b", "c", "d", "e"]);
 
     let expected_subsets = [vec![0], vec![1], vec![2], vec![3], vec![4]];
     verify_subsets(&disjoint_set, &expected_subsets);
@@ -31,7 +31,7 @@ fn constructor_prodcues_singletons() {
 #[test]
 fn constructor_constructs_correct_len_and_if_empty() {
     for size in 0..100 {
-        let disjoint_set = DisjointSet::new(size);
+        let disjoint_set = DisjointSetVec::from(vec![true; size]);
         assert_eq!(disjoint_set.len(), size);
         assert_eq!(disjoint_set.is_empty(), size == 0);
     }
@@ -40,7 +40,7 @@ fn constructor_constructs_correct_len_and_if_empty() {
 #[test]
 fn joining_and_is_joined_dont_change_len() {
     for size in 3..103 {
-        let mut disjoint_set = DisjointSet::new(size);
+        let mut disjoint_set = DisjointSetVec::from(vec![4.5; size]);
 
         disjoint_set.join(0, 1);
         assert!(!disjoint_set.is_joined(1, 2));
@@ -56,7 +56,7 @@ fn joining_and_is_joined_dont_change_len() {
 
 #[test]
 fn join_one_pair() {
-    let mut disjoint_set = DisjointSet::new(3);
+    let mut disjoint_set = DisjointSetVec::from(vec![3, 3, 4]);
     disjoint_set.join(0, 1);
     let expected_subsets = [vec![0, 1], vec![2]];
     verify_subsets(&disjoint_set, &expected_subsets);
@@ -64,7 +64,7 @@ fn join_one_pair() {
 
 #[test]
 fn join_element_with_itself() {
-    let mut disjoint_set = DisjointSet::new(3);
+    let mut disjoint_set = DisjointSetVec::from(vec![-1, -2, -3]);
     disjoint_set.join(1, 1);
     let expected_subsets = [vec![0], vec![1], vec![2]];
     verify_subsets(&disjoint_set, &expected_subsets);
@@ -72,7 +72,7 @@ fn join_element_with_itself() {
 
 #[test]
 fn join_one_pair_twice() {
-    let mut disjoint_set = DisjointSet::new(4);
+    let mut disjoint_set = DisjointSetVec::from(vec!["test"; 4]);
     disjoint_set.join(0, 1);
     disjoint_set.join(0, 1);
     let expected_subsets = [vec![0, 1], vec![2], vec![3]];
@@ -81,7 +81,7 @@ fn join_one_pair_twice() {
 
 #[test]
 fn join_one_pair_and_reversed() {
-    let mut disjoint_set = DisjointSet::new(2);
+    let mut disjoint_set = DisjointSetVec::from(vec![true, false]);
     disjoint_set.join(0, 1);
     disjoint_set.join(1, 0);
     let expected_subsets = [vec![0, 1]];
@@ -90,7 +90,7 @@ fn join_one_pair_and_reversed() {
 
 #[test]
 fn join_two_pairs_overlapping() {
-    let mut disjoint_set = DisjointSet::new(4);
+    let mut disjoint_set = DisjointSetVec::from(vec!["test"; 4]);
     disjoint_set.join(0, 1);
     disjoint_set.join(1, 3);
     let expected_subsets = [vec![0, 1, 3], vec![2]];
@@ -99,7 +99,7 @@ fn join_two_pairs_overlapping() {
 
 #[test]
 fn join_two_pairs_non_overlapping() {
-    let mut disjoint_set = DisjointSet::new(10);
+    let mut disjoint_set = DisjointSetVec::from(vec![3; 10]);
     disjoint_set.join(0, 4);
     disjoint_set.join(2, 7);
     let expected_subsets = [
@@ -117,7 +117,7 @@ fn join_two_pairs_non_overlapping() {
 
 #[test]
 fn join_two_non_trivial_subsets() {
-    let mut disjoint_set = DisjointSet::new(5);
+    let mut disjoint_set = DisjointSetVec::from(vec![3; 5]);
     disjoint_set.join(0, 1);
     disjoint_set.join(2, 3);
     let expected_subsets = [vec![0, 1], vec![2, 3], vec![4]];
@@ -130,7 +130,7 @@ fn join_two_non_trivial_subsets() {
 
 #[test]
 fn arbitrary_sequence_of_join_and_is_joined() {
-    let mut disjoint_set = DisjointSet::new(5);
+    let mut disjoint_set = DisjointSetVec::from(vec![3; 5]);
     assert!(!disjoint_set.is_joined(1, 2));
     assert!(!disjoint_set.is_joined(1, 2));
     assert!(disjoint_set.is_joined(1, 1));
@@ -175,49 +175,49 @@ fn arbitrary_sequence_of_join_and_is_joined() {
 #[test]
 #[should_panic]
 fn panic_join_first_element_oob() {
-    let mut disjoint_set = DisjointSet::new(1000);
+    let mut disjoint_set = DisjointSetVec::from(vec![3; 1000]);
     disjoint_set.join(1000, 5);
 }
 
 #[test]
 #[should_panic]
 fn panic_join_second_element_oob() {
-    let mut disjoint_set = DisjointSet::new(100);
+    let mut disjoint_set = DisjointSetVec::from(vec![3; 100]);
     disjoint_set.join(0, 1000000000000);
 }
 
 #[test]
 #[should_panic]
 fn panic_join_both_elements_oob() {
-    let mut disjoint_set = DisjointSet::new(0);
+    let mut disjoint_set = DisjointSetVec::from(vec![3; 0]);
     disjoint_set.join(0, 0);
 }
 
 #[test]
 #[should_panic]
 fn panic_is_joined_first_element_oob() {
-    let disjoint_set = DisjointSet::new(1000);
+    let disjoint_set = DisjointSetVec::from(vec![3; 1000]);
     let _ = disjoint_set.is_joined(1000, 5);
 }
 
 #[test]
 #[should_panic]
 fn panic_is_joined_second_element_oob() {
-    let disjoint_set = DisjointSet::new(100);
+    let disjoint_set = DisjointSetVec::from(vec![3; 100]);
     let _ = disjoint_set.is_joined(0, 1000000000000);
 }
 
 #[test]
 #[should_panic]
 fn panic_is_joined_both_elements_oob() {
-    let disjoint_set = DisjointSet::new(0);
+    let disjoint_set = DisjointSetVec::from(vec![3; 0]);
     let _ = disjoint_set.is_joined(0, 0);
 }
 
 #[test]
 #[should_panic]
 fn clone_clones() {
-    let mut disjoint_set = DisjointSet::new(5);
+    let mut disjoint_set = DisjointSetVec::from(vec![3; 5]);
     disjoint_set.join(2, 4);
     disjoint_set.join(3, 4);
 
@@ -231,16 +231,16 @@ fn clone_clones() {
 
 #[test]
 fn different_len_not_equal() {
-    let l = DisjointSet::new(5);
-    let r = DisjointSet::new(6);
+    let l = DisjointSetVec::from(vec![3; 5]);
+    let r = DisjointSetVec::from(vec![3; 6]);
 
     assert_ne!(l, r);
 }
 
 #[test]
 fn same_len_not_equal() {
-    let mut l = DisjointSet::new(5);
-    let mut r = DisjointSet::new(5);
+    let mut l = DisjointSetVec::from(vec![3; 5]);
+    let mut r = DisjointSetVec::from(vec![3; 5]);
 
     l.join(2, 4);
     r.join(2, 3);
@@ -250,8 +250,8 @@ fn same_len_not_equal() {
 
 #[test]
 fn different_joining_order_equal() {
-    let mut l = DisjointSet::new(5);
-    let mut r = DisjointSet::new(5);
+    let mut l = DisjointSetVec::from(vec![3; 5]);
+    let mut r = DisjointSetVec::from(vec![3; 5]);
 
     l.join(2, 4);
     l.join(1, 3);
@@ -268,9 +268,9 @@ fn different_joining_order_equal() {
 
 #[test]
 fn get_sets_empty() {
-    let ds = DisjointSet::new(0);
+    let ds = DisjointSetVec::from(vec![3; 0]);
 
-    let sets = ds.get_sets();
+    let sets = ds.get_index_sets();
 
     let expected: Vec<Vec<_>> = Vec::new();
 
@@ -279,9 +279,9 @@ fn get_sets_empty() {
 
 #[test]
 fn get_sets_singletons() {
-    let ds = DisjointSet::new(10);
+    let ds = DisjointSetVec::from(vec![3; 10]);
 
-    let sets = ds.get_sets();
+    let sets = ds.get_index_sets();
 
     let expected = vec![
         vec![0],
@@ -301,7 +301,7 @@ fn get_sets_singletons() {
 
 #[test]
 fn get_sets_all_in_one() {
-    let mut ds = DisjointSet::new(8);
+    let mut ds = DisjointSetVec::from(vec![3; 8]);
 
     ds.join(0, 1);
     ds.join(2, 3);
@@ -311,7 +311,7 @@ fn get_sets_all_in_one() {
     ds.join(5, 7);
     ds.join(3, 4);
 
-    let sets = ds.get_sets();
+    let sets = ds.get_index_sets();
 
     let expected = [vec![0, 1, 2, 3, 4, 5, 6, 7]];
 
@@ -320,14 +320,14 @@ fn get_sets_all_in_one() {
 
 #[test]
 fn get_sets_complex() {
-    let mut ds = DisjointSet::new(5);
+    let mut ds = DisjointSetVec::from(vec![3; 5]);
 
     ds.join(0, 3);
     ds.join(2, 1);
-    ds.add_singleton();
+    ds.push(100);
     ds.join(2, 5);
 
-    let sets = ds.get_sets();
+    let sets = ds.get_index_sets();
 
     let expected = [vec![0, 3], vec![1, 2, 5], vec![4]];
 
@@ -336,30 +336,30 @@ fn get_sets_complex() {
 
 #[test]
 fn construct_with_capacity() {
-    let ds = DisjointSet::with_capacity(30);
+    let ds: DisjointSetVec<&&Option<&&char>> = DisjointSetVec::with_capacity(30);
     assert!(ds.is_empty());
 }
 
 #[test]
 fn add_singleton_produces_singleton() {
-    let mut ds = DisjointSet::new(3);
+    let mut ds = DisjointSetVec::from(vec![-1, -2, -3]);
     ds.join(0, 2);
     verify_subsets(&ds, &[vec![0, 2], vec![1]]);
-    ds.add_singleton();
+    ds.push(0);
     verify_subsets(&ds, &[vec![0, 2], vec![1], vec![3]]);
-    ds.add_singleton();
+    ds.push(0);
     verify_subsets(&ds, &[vec![0, 2], vec![1], vec![3], vec![4]]);
 }
 
 #[test]
 fn can_join_elements_added_later() {
-    let mut ds = DisjointSet::new(3);
+    let mut ds = DisjointSetVec::from(vec![-1, -2, -3]);
 
     ds.join(0, 2);
 
-    ds.add_singleton();
-    ds.add_singleton();
-    ds.add_singleton();
+    ds.push(3);
+    ds.push(4);
+    ds.push(-33);
 
     ds.join(4, 5);
     ds.join(3, 0);
@@ -369,7 +369,168 @@ fn can_join_elements_added_later() {
 }
 
 #[test]
+fn get_works_in_all_circumstances() {
+    let mut ds = DisjointSetVec::from(vec!["a", "a", "b"]);
+
+    assert_eq!(ds.get(0), Some(&"a"));
+    assert_eq!(ds.get(1), Some(&"a"));
+    assert_eq!(ds.get(2), Some(&"b"));
+    assert_eq!(ds.get(3), None);
+    assert_eq!(ds.get(4), None);
+
+    ds.join(0, 2);
+
+    assert_eq!(ds.get(0), Some(&"a"));
+    assert_eq!(ds.get(1), Some(&"a"));
+    assert_eq!(ds.get(2), Some(&"b"));
+    assert_eq!(ds.get(3), None);
+    assert_eq!(ds.get(4), None);
+
+    ds[0] = "c";
+
+    assert_eq!(ds.get(0), Some(&"c"));
+    assert_eq!(ds.get(1), Some(&"a"));
+    assert_eq!(ds.get(2), Some(&"b"));
+    assert_eq!(ds.get(3), None);
+    assert_eq!(ds.get(4), None);
+
+    ds.push("d");
+
+    assert_eq!(ds.get(0), Some(&"c"));
+    assert_eq!(ds.get(1), Some(&"a"));
+    assert_eq!(ds.get(2), Some(&"b"));
+    assert_eq!(ds.get(3), Some(&"d"));
+    assert_eq!(ds.get(4), None);
+}
+
+#[test]
+fn iter_works() {
+    let mut ds = DisjointSetVec::from(vec!["a", "a", "b"]);
+
+    ds.join(0, 2);
+    ds.push("d");
+
+    let expected = ["a", "a", "b", "d"];
+
+    assert!(ds.iter().eq(expected.iter()));
+}
+
+#[test]
+fn iter_mut_works() {
+    let mut ds = DisjointSetVec::from(vec![3, 4, 5]);
+
+    ds.join(0, 2);
+    ds.push(0);
+
+    let clone = ds.clone();
+
+    for val in ds.iter_mut() {
+        *val *= 2;
+    }
+
+    let mut expected_ds = DisjointSetVec::from(vec![6, 8, 10, 0]);
+    expected_ds.join(0, 2);
+
+    assert_eq!(ds, expected_ds);
+    assert_ne!(ds, clone);
+}
+
+#[test]
+fn indexing_works() {
+    let mut ds = DisjointSetVec::from(vec![3, 4, 5]);
+
+    ds.join(0, 2);
+    ds.push(0);
+
+    assert_eq!(ds[0], 3);
+    assert_eq!(ds[1], 4);
+    assert_eq!(ds[2], 5);
+    assert_eq!(ds[3], 0);
+}
+
+#[test]
 #[should_panic]
-fn with_capacity_panics() {
-    let _ = DisjointSet::with_capacity(isize::MAX as usize - 1);
+fn indexing_panics() {
+    let ds = DisjointSetVec::from(vec![3, 4, 5]);
+    let _ = ds[3];
+}
+
+#[test]
+fn mutable_indexing_works() {
+    let mut ds = DisjointSetVec::from(vec![3, 4, 5]);
+
+    ds.join(0, 2);
+    ds.push(0);
+
+    ds[1] = 33;
+
+    assert_eq!(ds[0], 3);
+    assert_eq!(ds[1], 33);
+    assert_eq!(ds[2], 5);
+    assert_eq!(ds[3], 0);
+}
+
+#[test]
+#[should_panic]
+fn mutable_indexing_panics() {
+    let mut ds = DisjointSetVec::from(vec![3, 4, 5]);
+    ds[3] = 3;
+}
+
+#[test]
+fn into_iter_works() {
+    let expected = vec![3, 4, 5];
+
+    let ds = DisjointSetVec::from(expected.clone());
+
+    let mut vec = Vec::with_capacity(3);
+
+    for val in ds {
+        vec.push(val);
+    }
+
+    assert_eq!(vec, expected);
+}
+
+#[test]
+fn ref_into_iter_works() {
+    let expected = vec![3, 4, 5];
+
+    let ds = DisjointSetVec::from(expected.clone());
+
+    let mut vec = Vec::with_capacity(3);
+
+    for &val in &ds {
+        vec.push(val);
+    }
+
+    assert_eq!(vec, expected);
+}
+
+#[test]
+fn mut_ref_into_iter_works() {
+    let mut ds = DisjointSetVec::from(vec![2, 3]);
+
+    for val in &mut ds {
+        *val += 10;
+    }
+
+    assert_eq!(ds.get(0), Some(&12));
+    assert_eq!(ds.get(1), Some(&13));
+    assert_eq!(ds.get(2), None);
+}
+
+#[test]
+fn different_ways_of_empty_construction() {
+    let empty_from: DisjointSetVec<bool> = DisjointSetVec::from(vec![]);
+    assert!(empty_from.is_empty());
+
+    let empty_new = DisjointSetVec::new();
+    assert!(empty_new.is_empty());
+
+    let empty_default = DisjointSetVec::default();
+    assert!(empty_new.is_empty());
+
+    assert_eq!(empty_from, empty_new);
+    assert_eq!(empty_from, empty_default);
 }