+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
-## v0.1.0 - 2021-04-12
+## [v0.2.0] - 2021-04-13
+
+### Added
+- Implemented `Debug`, `Clone`, `PartialEq` and `Default` for `DisjointSet`.
+- Added `with_capacity`, `add_singleton`, `get_sets` to `DisjointSet`.  
+- Added `DisjointSetVec`.
+
+## [v0.1.0] - 2021-04-12
 
 - Initial release
+
+[v0.1.0]: https://github.com/jogru0/disjoint/commit/15bb8dce2a5f33812fe237d19354a792612fd92c
+[v0.2.0]: https://github.com/jogru0/disjoint/compare/v0.1.0...v0.2.0